conversation/bigquery-sql-agent/requirements.txt
functions-framework==3.*
google-cloud-bigquery 
google-cloud-aiplatform
pandas 
db-dtypes
tabulate
mesop